Fix for update book, during edit to avoid duplicate isbn check
During the edit to update the existing book details, the code previously included a check to validate again with ISBN numbers, which caused a duplicate ISBN issue. 
Fixed code such that during the update process, the book's isbn not need to be rechecked. It is only needed when adding a new book